refactored services structure in duty app
