+++ core/apps/duties/services/__init__.py
@@ -1,0 +1,5 @@
+from .base import *  # noqa: F403
+from .kitchen_duty import *  # noqa: F403
+from .swap_duties import *  # noqa: F403
+from .swap_people import *  # noqa: F403
+from .utils import *  # noqa: F403
